get rid of the useless test for insert

+++ test/category-test.php
@@ -52,7 +52,7 @@ class CategoryTest extends UnitTestCase {
 	 * tears down the connection to mysql and deletes the test instance object
 	 **/
 	public function tearDown() {
-		if($this->category !== null) {
+		if($this->category !== null && $this->category->getCategoryId() !== null) {
 			$this->category->delete($this->mysqli);
 			$this->category = null;
 		}
